fix (#5): removed webpack and updated project structure
